Minor consistency updates
The method that returns the size of the binary search tree used to be
called "size" (as opposed to "getSize"). This was to maintain
consistency with the "size" method of Java collections. Since
BinarySearchTree does not implement the Collection interface, I
eventually decided to change "size" to "getSize" to maintain consistency
with the rest of the class.